Add files via upload
In order to access data from the bulk of people anytime anywhere we are creating our own server using nodemcu. we will provide user an ip-address so once the address is open user will be directed to a site for details login.Final goal of the task is to sent the emails to the user using python program.